added all changes except demand

+++ header/Producer.h
@@ -30,7 +30,7 @@ class Producer : public Firm {
   private:
     std::unordered_map<Firm *, Plan *> customer_to_draft_plan;
 
-    int get_max_order_quantity(Product * product);
+    double get_max_order_quantity(Product * product);
     void log_draft_plan(const Plan * draft_plan);
     void log_dropped_order(const Order * order);
 };
